+++ src/Components/Navbars/Bottom Navs/bottomNav.css
@@ -7,6 +7,7 @@
   position: fixed;
   bottom: 0;
   width: 100%;
+  z-index: 5;
 }
 .bottomNavContainerItems {
   font-size: 1.2rem;
