+++ src/Components/Tweet/CreateTweet.jsx
@@ -1,19 +1,6 @@
-import { EditorState } from "draft-js";
-import Editor from "@draft-js-plugins/editor";
-import createHashtagPlugin from "@draft-js-plugins/hashtag";
-
-import { useState } from "react";
-
+import TweetEditor from "./Tweet Editor/TweetEditor";
 import styles from "./createTweet.module.css";
-import "draft-js/dist/Draft.css";
-
-const hashtagPlugin = createHashtagPlugin({ theme: styles });
-
 const CreateTweet = () => {
-  const [editorState, setEditorState] = useState(() =>
-    EditorState.createEmpty()
-  );
-
   return (
     <div className={styles.createTweetContainer}>
       <img
@@ -21,14 +8,7 @@ const CreateTweet = () => {
         src="https://avatars.dicebear.com/api/micah/alex.svg"
         alt="User Avatar"
       />
-      <div className={styles.editorStyle}>
-        <Editor
-          editorState={editorState}
-          onChange={setEditorState}
-          placeholder="What's happening?"
-          plugins={[hashtagPlugin]}
-        />
-      </div>
+      <TweetEditor />
     </div>
   );
 };
