+++ src/Components/Tweet/Tweet Editor/tweetEditor.module.css
@@ -1,0 +1,20 @@
+.editorStyle {
+  /* background: rgb(128, 28, 28); */
+  color: rgb(190, 20, 20);
+  width: 80%;
+  display: flex;
+  flex-direction: column;
+}
+.hashtag {
+  color: #ee7320;
+}
+.tweetButton {
+  border: none;
+  width: 4rem;
+  height: 1.8rem;
+  border-radius: 20px;
+  background: rgb(4, 100, 226);
+  color: rgb(247, 245, 241);
+  cursor: pointer;
+  align-self: flex-end;
+}
