+++ src/Components/Tweet/createTweet.module.css
@@ -14,11 +14,3 @@
   height: 2.5rem;
   align-self: flex-start;
 }
-.editorStyle {
-  /* background: rgb(128, 28, 28); */
-  color: rgb(190, 20, 20);
-  width: 80%;
-}
-.hashtag {
-  color: #ee7320;
-}
